chore(BUX-686): where_mocks transfered to another file
